Reassign OW IDs

+++ src/frontier_util.c
@@ -665,7 +665,7 @@ static const struct WindowTemplate sRankingHallRecordsWindowTemplate =
 };
 
 // Second field - whether the character is female.
-static const u8 sFrontierBrainObjEventGfx[NUM_FRONTIER_FACILITIES][2] =
+static const u16 sFrontierBrainObjEventGfx[NUM_FRONTIER_FACILITIES][2] =
 {
     [FRONTIER_FACILITY_TOWER]   = {OBJ_EVENT_GFX_ANABEL,  TRUE},
     [FRONTIER_FACILITY_DOME]    = {OBJ_EVENT_GFX_TUCKER,  FALSE},
